# [benchmarks] Use std.mem.doNotOptimizeAway to avoid data collisions