@jcollie 2025-09-06
> shows the only difference is the offsets, which somehow have a large impact on the result of the benchmark.

Perhaps because the array isn't aligned to the CPU's natural fetch boundaries? Maybe if you force the alignment of `buf` to `usize` it would eliminate the difference?

```
>>> 0x10001728c % 8
4
>>> 0x1000176d8 % 8
0
```